# src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TCF API",
    "summary": "API to calculate the predicted price for a vehicle with given features",
    "description": "\n## How to use this API\nlore ipsum dolor sit amet consectetur adipiscing elit\n\n## Available ENUMs\nlore ipsum dolor sit amet consectetur adipiscing elit\n",
    "contact": {
      "name": "Vadym Tatarinov",
      "url": "https://ford-parts.com.ru/",
      "email": "utikpuhlik@protonmail.ch"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local environment"
    },
    {
      "url": "https://api.dev.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Development environment"
    },
    {
      "url": "https://api.stage.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Staging environment"
    },
    {
      "url": "https://api.prod.alpha-analytics.dev/v2/pricingv2/predict_price",
      "description": "Production environment"
    }
  ],
  "paths": {
...
    "/waybills": {
      "get": {
        "tags": [
          "Waybills"
        ],
        "summary": "Get Waybills",
        "description": "Get waybills by type and user ID",
        "operationId": "get_waybills_waybills_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "WAYBILL_IN",
                    "WAYBILL_OUT",
                    "WAYBILL_RETURN"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Waybill Type"
            }
          },
          {
            "name": "is_pending",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Pending"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "search_term",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search Term"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WaybillSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Waybills"
        ],
        "summary": "Create Waybill",
        "operationId": "create_waybill_waybills_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaybillPostSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaybillSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/waybills/{waybill_id}": {
      "get": {
        "tags": [
          "Waybills"
        ],
        "summary": "Get Waybill",
        "description": "Get waybill by ID",
        "operationId": "get_waybill_waybills__waybill_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waybill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaybillSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Waybills"
        ],
        "summary": "Delete Waybill",
        "operationId": "delete_waybill_waybills__waybill_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waybill Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/waybills/{waybill_id}/offers": {
      "get": {
        "tags": [
          "Waybills"
        ],
        "summary": "Get Waybill Offers",
        "description": "Get all offers in waybill",
        "operationId": "get_waybill_offers_waybills__waybill_id__offers_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waybill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaybillOfferSchema"
                  },
                  "title": "Response Get Waybill Offers Waybills  Waybill Id  Offers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Waybills"
        ],
        "summary": "Add Offer To Waybill",
        "operationId": "add_offer_to_waybill_waybills__waybill_id__offers_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waybill Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaybillOfferPostSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaybillOfferSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/waybills/{waybill_id}/commit": {
      "post": {
        "tags": [
          "Waybills"
        ],
        "summary": "Commit Waybill",
        "operationId": "commit_waybill_waybills__waybill_id__commit_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "waybill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Waybill Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaybillSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_WaybillSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WaybillSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "size"
        ],
        "title": "Page[WaybillSchema]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaybillOfferPostSchema": {
        "properties": {
          "offer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Offer Id"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Quantity",
            "examples": [10]
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "examples": [
              "BSG"
            ]
          },
          "manufacturer_number": {
            "type": "string",
            "minLength": 1,
            "title": "Manufacturer Number",
            "examples": [
              "BSG-12345"
            ]
          },
          "price_rub": {
            "type": "number",
            "title": "Price Rub",
            "examples": [1000]
          }
        },
        "type": "object",
        "required": [
          "offer_id",
          "quantity",
          "brand",
          "manufacturer_number",
          "price_rub"
        ],
        "title": "WaybillOfferPostSchema"
      },
      "WaybillOfferSchema": {
        "properties": {
          "offer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Offer Id"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Quantity",
            "examples": [10]
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "examples": [
              "BSG"
            ]
          },
          "manufacturer_number": {
            "type": "string",
            "minLength": 1,
            "title": "Manufacturer Number",
            "examples": [
              "BSG-12345"
            ]
          },
          "price_rub": {
            "type": "number",
            "title": "Price Rub",
            "examples": [1000]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "waybill_id": {
            "type": "string",
            "format": "uuid",
            "title": "Waybill Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "address_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Id",
            "examples": [
              "AA-TEST"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "category_slug": {
            "type": "string",
            "title": "Category Slug",
            "examples": [
              "svechi-ford"
            ]
          },
          "category_name": {
            "type": "string",
            "title": "Category Name",
            "examples": [
              "Свечи"
            ]
          },
          "sub_category_slug": {
            "type": "string",
            "title": "Sub Category Slug",
            "examples": [
              "svechi-zazhiganiia"
            ]
          },
          "sub_category_name": {
            "type": "string",
            "title": "Sub Category Name",
            "examples": [
              "Свечи зажигания"
            ]
          }
        },
        "type": "object",
        "required": [
          "offer_id",
          "quantity",
          "brand",
          "manufacturer_number",
          "price_rub",
          "id",
          "waybill_id",
          "product_name",
          "image_url",
          "category_slug",
          "category_name",
          "sub_category_slug",
          "sub_category_name"
        ],
        "title": "WaybillOfferSchema"
      },
      "WaybillPostSchema": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "waybill_type": {
            "type": "string",
            "enum": [
              "WAYBILL_IN",
              "WAYBILL_OUT",
              "WAYBILL_RETURN"
            ],
            "title": "Waybill Type"
          },
          "is_pending": {
            "type": "boolean",
            "title": "Is Pending",
            "examples": [true]
          },
          "counterparty_name": {
            "type": "string",
            "title": "Counterparty Name",
            "examples": [
              "ООО Рога и Копыта"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "waybill_type",
          "is_pending",
          "counterparty_name"
        ],
        "title": "WaybillPostSchema"
      },
      "WaybillSchema": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "waybill_type": {
            "type": "string",
            "enum": [
              "WAYBILL_IN",
              "WAYBILL_OUT",
              "WAYBILL_RETURN"
            ],
            "title": "Waybill Type"
          },
          "is_pending": {
            "type": "boolean",
            "title": "Is Pending",
            "examples": [true]
          },
          "counterparty_name": {
            "type": "string",
            "title": "Counterparty Name",
            "examples": [
              "ООО Рога и Копыта"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "examples": [
              "Vasiliy Pinov"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "waybill_type",
          "is_pending",
          "counterparty_name",
          "id",
          "author",
          "created_at",
          "updated_at"
        ],
        "title": "WaybillSchema"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}